add-feed

+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -2,9 +2,12 @@ package ru.yandex.practicum.filmorate.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.Genre;
 import ru.yandex.practicum.filmorate.storage.film.FilmStorage;
+import ru.yandex.practicum.filmorate.utility.EventOperation;
+import ru.yandex.practicum.filmorate.utility.EventType;
 
 import java.util.Collection;
 import java.util.Comparator;
@@ -17,9 +20,12 @@ public class FilmService {
 
     private final FilmStorage filmDbStorage;
 
+    private final EventService eventService;
+
     @Autowired
-    public FilmService(FilmStorage filmDbStorage) {
+    public FilmService(FilmStorage filmDbStorage, EventService eventService) {
         this.filmDbStorage = filmDbStorage;
+        this.eventService = eventService;
     }
 
     public Optional<Film> postFilm(Film film) {
@@ -49,10 +55,15 @@ public class FilmService {
     }
 
     public Optional<Film> putLikeToFilm(int filmId, int userId) {
+        eventService.createEvent(userId, EventType.LIKE, EventOperation.ADD, filmId);
         return filmDbStorage.putLikeToFilm(filmId, userId);
     }
 
     public Optional<Film> deleteLikeToFilm(int filmId, int userId) {
+        if (userId < 0) {
+            throw new NotFoundException("Не может быть отрицательного id");
+        }
+        eventService.createEvent(userId, EventType.LIKE, EventOperation.REMOVE, filmId);
         return filmDbStorage.deleteLikeFromFilm(filmId, userId);
     }
 
@@ -63,5 +74,4 @@ public class FilmService {
                 .limit(count)
                 .collect(Collectors.toList());
     }
-
 }
